fix: enforce per-device correctness across both backends (multi-GPU Phase 0)
Phase 0 of docs/dev/multi_gpu_plan.md: make every native entry point
correct when the ambient CUDA device differs from the scene/input device,
with zero behavioral change for valid single-device inputs.

Torch backend:
- Add c10::cuda::CUDAGuard at every op entry that launches kernels or CUB
  (scene, edge, reflection, diffraction, sdf, penetration, stats, and all
  RF scattering/transmission/layer-stack/wedge families).
- Add a same-device-index check (require_same_device) and apply it to the
  RF families that previously only checked is_cuda().
- coop_launch_config_for_device now queries occupancy under a guard for
  the requested device instead of the ambient one.
- native_compat.h stream access and the reflection EPC/dedup launch
  helpers are now device-explicit; compute_edge_optix_aabbs_cuda derives
  its device from the buffers it operates on.
- penetration ops resolve the OptiX context/pipeline from the validated
  scene device instead of the ambient context.
- stats.cu: use int64_t instead of long long (data_ptr<long long> only
  links on Windows; Linux portability fix, no-op on Windows).

Dr.Jit backend:
- OptixLaunchPipeline records its build-time device and launches assert
  it; Scene records its build device and every public query asserts the
  current device matches, failing loudly instead of corrupting state
  after set_device().

ADR-0026: refresh the 11 pinned RF source hashes (ADR text + guard test)
as an intentional identity update; the change is one CUDAGuard per host
entry, with kernels, launch counts, flags, reduction order, and AD
semantics untouched.

Tests: extend cpp/scattering_test.cpp cross-device rejection to all six
families; add a 2-GPU ambient-device-independence smoke suite
(test_multi_device_smoke.py, 10 tests) and a Dr.Jit device-binding test.
Verified on 2x RTX A6000: full Torch suite 198 tests OK (19 skipped).

+++ backends/torch/src/torch_ext/common/stats.cu
@@ -12,7 +12,7 @@ __global__ void reflection_trace_stats_kernel(
     const float *__restrict__ t,
     int64_t ray_count,
     int64_t max_bounces,
-    long long *__restrict__ counts,
+    int64_t *__restrict__ counts,
     double *__restrict__ checksum) {
     const int ray_idx = blockIdx.x * blockDim.x + threadIdx.x;
     if (ray_idx >= ray_count)
@@ -42,10 +42,10 @@ __global__ void diffraction_path_stats_kernel(
     const bool *__restrict__ valid,
     const float *__restrict__ delay,
     int64_t capacity,
-    long long *__restrict__ valid_count,
+    int64_t *__restrict__ valid_count,
     double *__restrict__ checksum) {
     if (blockIdx.x == 0 && threadIdx.x == 0) {
-        valid_count[0] = static_cast<long long>(count[0]);
+        valid_count[0] = static_cast<int64_t>(count[0]);
     }
     const int path_idx = blockIdx.x * blockDim.x + threadIdx.x;
     if (path_idx >= capacity)
@@ -119,7 +119,7 @@ std::tuple<at::Tensor, at::Tensor> reflection_trace_stats_cuda(
         t.data_ptr<float>(),
         ray_count,
         max_bounces,
-        counts.data_ptr<long long>(),
+        counts.data_ptr<int64_t>(),
         checksum.data_ptr<double>());
     return std::make_tuple(counts, checksum);
 }
@@ -141,7 +141,7 @@ std::tuple<at::Tensor, at::Tensor> diffraction_path_stats_cuda(
         valid.data_ptr<bool>(),
         delay.data_ptr<float>(),
         capacity,
-        valid_count.data_ptr<long long>(),
+        valid_count.data_ptr<int64_t>(),
         checksum.data_ptr<double>());
     return std::make_tuple(valid_count, checksum);
 }
